pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline-migration-tool"
requires-python = ">=3.12"
authors = [
  {name = "Konflux Build Team"},
]
description = "Pipeline migration tool for Konflux CI."
readme = "README.md"
license = {text = "Apache-2.0"}
keywords = ["pipeline", "migration", "konflux", "ci"]
classifiers = [
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Build Tools",
]
dynamic = ["version"]

dependencies = [
  "requests==2.32.3",
  "ruamel.yaml==0.18.6",
  "oras==0.2.25",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-cov",
  "responses",
  "tox",
]

[project.urls]
Homepage = "https://github.com/konflux-ci/pipeline-migration-tool"
Documentation = "https://github.com/konflux-ci/pipeline-migration-tool/blob/main/README.md"
Issues = "https://github.com/konflux-ci/pipeline-migration-tool/issues"

[project.scripts]
pipeline-migration-tool = "pipeline_migration.cli:entry_point"

[tool.setuptools.dynamic]
version = {attr = "pipeline_migration.__version__"}

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
skip_covered = true
omit = [
  "tests/*"
]
